requirements.txt
openai
python-dotenv
streamlit 